+++ scripts/Groovy/lib/entities/Query.groovy
@@ -5,6 +5,7 @@ import groovy.transform.ToString
 @ToString(includeNames = true, ignoreNulls = true)
 class Query {
     List<String> select
+    String from
     List<Reduction> reductions
     CompoundFilter where
 }