+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2024-03-24
+### Added
+- `from` clause to the query in `CsvQuery`
+
 ## 2024-03-23
 ### Changed
 - `CsvQuery` to improve speed through asynchronous processing